requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
websockets==12.0
bcrypt==4.0.1
pydantic==2.5.2
alembic==1.12.1
python-socketio==5.9.0
python-engineio==4.6.1
